requirements.txt
fastapi==0.78.0
sqlalchemy==1.4.40
sqlalchemy-asyncio==0.19.0
bcrypt==3.2.0
pydantic==1.9.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
uvicorn==0.18.2
python-dotenv==0.20.0
asyncpg==0.26.0
streamlit==1.12.0
requests==2.28.1
transformers==4.21.1
langchain-openai==0.3.5
langchain-chroma==0.1.2
langchain==0.5.5
